README: improve example
The example in the README file and the related example files are not
consistent. Change the "wag200g" device for a fake board and switch from
an "Example" manufacturer.

Unfortunately Github does not allow yet the usage of AsciiDoc include
directives for security reasons. Use relative links for the moment.