docs: add Roff manpage
This commit makes use of the asciidoctor processor (distribution package
of the same name) to convert the AsciiDoc manpage into the Roff format.

It can be generated from the command line with:

    $ asciidoctor -b manpage docs/dsatest.1.adoc

But to simplify this and avoid the inclusion of a custom script, provide
a Rakefile (Ruby equivalent to a Makefile) to use directory the Ruby
library of the asciidoctor package.

The docs/dsatest.1 document is the result of running `rake` in docs/.