docs: prepare website
This commit makes use of the AsciiDoc preprocessor directives to fine
tune the links (*) and documents structure for the generated HTML pages.

For the moment, the website will simply be the README file tweaked into
the HTML format. The Rakefile has been updated to render it, as well as
the HTML version of the manpage and the example bench.

In order to host the website on Github, the Github Pages feature of the
repository has to be configured to render the docs/ directory of the
master branch.

(*) Github defines the "env-github" attribute when converting AsciiDoc
file. This makes it possible to distinct whether a file was rendered
from the Github repository or from us.